requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
yt-dlp
requests==2.31.0
gunicorn==21.2.0
urllib3==2.0.7

